requirements.txt
django===4.2.2
pytest-django==4.5.1
factory-boy==3.2.1
django-environ==0.10.0
psycopg2-binary~=2.9.5
djangorestframework==3.14.0
pytest-dotenv==0.5.2
